refactor: restructure Trash module to views-based architecture
Move all Trash-related code to src/views/Trash following the new project structure:

- Move dialog components to src/views/Trash/components/
- ClearTrashDialog, DeleteItemsDialog
- StopSharingAndMoveToTrashDialog, StopSharingAndMoveToTrashDialogWrapper
- Move custom hooks to src/views/Trash/hooks/
- useTrashPagination and tests
- Move services to src/views/Trash/services/
- batch-processor, clear-trash, delete-items
- get_trash, move-items-to-trash and all tests
- Update all import paths across the codebase
