+++ src/app/drive/components/DriveExplorer/DriveExplorer.tsx
@@ -18,7 +18,7 @@ import DriveExplorerGrid from './DriveExplorerGrid/DriveExplorerGrid';
 import DriveExplorerList from './DriveExplorerList/DriveExplorerList';
 
 import { useHotkeys } from 'react-hotkeys-hook';
-import moveItemsToTrash from '../../../../services/Trash/move-items-to-trash';
+import moveItemsToTrash from '../../../../views/Trash/services/move-items-to-trash';
 
 import { Role } from '@internxt/sdk/dist/drive/share/types';
 import { WorkspaceData } from '@internxt/sdk/dist/workspaces';
@@ -29,10 +29,10 @@ import deviceService from '../../../core/services/device.service';
 import errorService from '../../../core/services/error.service';
 import navigationService from '../../../core/services/navigation.service';
 import RealtimeService, { SOCKET_EVENTS } from '../../../core/services/socket.service';
-import ClearTrashDialog from '../../../../views/Trash/ClearTrashDialog';
+import ClearTrashDialog from '../../../../views/Trash/components/ClearTrashDialog';
 import CreateFolderDialog from '../../../drive/components/CreateFolderDialog/CreateFolderDialog';
-import DeleteItemsDialog from '../../../../views/Trash/DeleteItemsDialog';
-import { useTrashPagination } from '../../../../hooks/useTrashPagination';
+import DeleteItemsDialog from '../../../../views/Trash/components/DeleteItemsDialog';
+import { useTrashPagination } from '../../../../views/Trash/hooks/useTrashPagination';
 import { useTranslationContext } from '../../../i18n/provider/TranslationProvider';
 import { uploadFoldersWithManager } from '../../../network/UploadFolderManager';
 import notificationsService, { ToastType } from '../../../notifications/services/notifications.service';
@@ -61,7 +61,7 @@ import ItemDetailsDialog from '../ItemDetailsDialog/ItemDetailsDialog';
 import MoveItemsDialog from '../MoveItemsDialog/MoveItemsDialog';
 import NameCollisionContainer from '../NameCollisionDialog/NameCollisionContainer';
 import ShareDialog from '../ShareDialog/ShareDialog';
-import StopSharingAndMoveToTrashDialogWrapper from '../../../../views/Trash/StopSharingAndMoveToTrashDialogWrapper';
+import StopSharingAndMoveToTrashDialogWrapper from '../../../../views/Trash/components/StopSharingAndMoveToTrashDialogWrapper';
 import UploadItemsFailsDialog from '../UploadItemsFailsDialog/UploadItemsFailsDialog';
 import WarningMessageWrapper from '../WarningMessage/WarningMessageWrapper';
 import './DriveExplorer.scss';
