+++ src/views/Trash/components/DeleteItemsDialog.tsx
@@ -6,7 +6,7 @@ import { setItemsToDelete } from 'app/store/slices/storage';
 import { useAppDispatch, useAppSelector } from 'app/store/hooks';
 import { RootState } from 'app/store';
 import { DriveItemData } from 'app/drive/types';
-import deleteItems from '../../services/Trash/delete-items';
+import deleteItems from '../services/delete-items';
 import { Button, Modal } from '@internxt/ui';
 import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
 import { planThunks } from 'app/store/slices/plan';
