+++ src/views/Trash/components/StopSharingAndMoveToTrashDialog.tsx
