+++ src/views/Trash/components/StopSharingAndMoveToTrashDialogWrapper.tsx
