+++ src/views/Trash/services/batch-processor.ts
