+++ src/views/Trash/services/clear-trash.test.ts
