+++ src/views/Trash/services/clear-trash.ts
